Adjusted duration variable type
It seems that it goes from maybe 0 to 1

+++ RosaServer/structs.h
@@ -388,7 +388,7 @@ struct Human {
 	padding unk30_1[0x35b4 - 0x355c - 4];
 	int progressBar;  // 35b4
 	int inventoryAnimationFlags; // 35b8
-	int inventoryAnimationDuration; // 35c0
+	float inventoryAnimationDuration; // 35c0
 	padding unk31[0x3758 - 0x35c0 - 4];
 	int gender;             // 3758
 	int head;               // 375c
